feat: support GitHub OAuth token as an app parameter (#10)
This allows an OAuth token to be provided to the app for use when
writing PR comments. CodeBuild does not always have a GitHub OAuth
token, e.g., when setup to build public repositories (no token required
to read from the repo). If provided, the app will store the token in AWS
Secrets Manager and use it when writing to GitHub.